backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.16
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
psycopg[binaary]